add env vars to control retry, and log polymarket response
