+++ packages/monitor-v2/src/monitor-polymarket/MonitorProposalsOrderBook.ts
@@ -17,8 +17,8 @@ import {
   getPolymarketOrderBook,
   getPolymarketProposedPriceRequestsOO,
   storeNotifiedProposals,
+  retryAsync,
 } from "./common";
-import { retryAsync } from "../utils/helpers";
 
 export async function monitorTransactionsProposedOrderBook(
   logger: typeof Logger,
@@ -52,7 +52,11 @@ async function processMarketProposal(market: OptimisticPriceRequest, params: Mon
   const networker = new Networker(logger);
   const questionID = calculatePolymarketQuestionID(market.ancillaryData);
   // set this to retry twice and wait 5 seconds between retries.
-  const polymarketInfo = await retryAsync(() => getPolymarketMarketInformation(params, questionID), 2, 5000);
+  const polymarketInfo = await retryAsync(
+    () => getPolymarketMarketInformation(logger, params, questionID),
+    params.retryAttempts,
+    params.retryDelayMs
+  );
   const orderBook = await getPolymarketOrderBook(params, polymarketInfo.clobTokenIds, networker);
   const orderFilledEvents = await getOrderFilledEvents(
     params,
