+++ packages/monitor-v2/src/monitor-polymarket/README.md
@@ -25,3 +25,5 @@ All the configuration should be provided with following environment variables:
 - `THRESHOLD_BIDS` Price threshold for loser outcome bids in the orderbook that triggers a notification (defaults to `0`).
 - `THRESHOLD_VOLUME` Volume threshold for the market that triggers a notification (defaults to `500000`).
 - `UNKNOWN_PROPOSAL_NOTIFICATION_INTERVAL` is the interval in seconds we allow to pass before notifying about a proposal event not found (defaults to `300`).
+- `RETRY_ATTEMPTS` optionally specify a number to retry various external queries to polymarket. Defaults to off.
+- `RETRY_DELAY_MS` optionally specify a delay in milliseconds between retries, defaults to 0 delay.
